requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.11.0
addict=2.4.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
aom=3.6.0=h313beb8_0
asitop=0.0.24=pypi_0
attrs=24.2.0=pypi_0
blas=1.0=openblas
blessed=1.20.0=pypi_0
blinker=1.9.0=pypi_0
bpy=4.0.0=pypi_0
brotli-python=1.0.9=py310h313beb8_8
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.9.24=hca03da5_0
cairo=1.16.0=h302bd0f_5
certifi=2024.8.30=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=pypi_0
configargparse=1.7=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.11=pypi_0
dash=2.18.2=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
dashing=0.1.0=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
dav1d=1.2.1=h80987f9_0
diffusers=0.31.0=pypi_0
docker-pycreds=0.4.0=pypi_0
expat=2.6.3=h313beb8_0
fastjsonschema=2.20.0=pypi_0
ffmpeg=6.1.1=h4beb5bb_0
filelock=3.13.1=py310hca03da5_0
flask=3.0.3=pypi_0
fontconfig=2.14.1=hee714a5_2
fonttools=4.55.0=pypi_0
freetype=2.12.1=h1192e45_0
fribidi=1.0.10=h1a28f6b_0
fsspec=2024.10.0=pypi_0
gdk-pixbuf=2.42.10=h80987f9_1
gettext=0.21.0=h13f89a0_1
giflib=5.2.2=h80987f9_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
glib=2.78.4=h313beb8_0
glib-tools=2.78.4=h313beb8_0
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py310h8c48613_0
graphite2=1.3.14=hc377ac9_1
harfbuzz=4.3.0=he9eebac_2
huggingface-hub=0.26.2=pypi_0
hydra-core=1.3.2=pypi_0
icu=73.1=h313beb8_0
idna=3.7=py310hca03da5_0
importlib-metadata=8.5.0=pypi_0
iopath=0.1.10=pyhd8ed1ab_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=py310hca03da5_1
joblib=1.4.2=pypi_0
jpeg=9e=h80987f9_3
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-core=5.7.2=pypi_0
kaleido-core=0.2.1=h27ca646_0
kiwisolver=1.4.7=pypi_0
lame=3.100=h1a28f6b_0
lcms2=2.12=hba8e193_0
leptonica=1.82.0=h18c3764_2
lerc=3.0=hc377ac9_0
libarchive=3.6.2=h62fee54_3
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libglib=2.78.4=h0a96307_0
libiconv=1.16=h80987f9_3
libjpeg-turbo=2.0.0=h1a28f6b_0
libogg=1.3.5=h1a28f6b_1
libopenblas=0.3.21=h269037a_0
libopus=1.3.1=h80987f9_1
libpng=1.6.39=h80987f9_0
librsvg=2.54.4=hb3bd4c3_3
libtheora=1.1.1=h1a28f6b_3
libtiff=4.5.1=h313beb8_0
libvorbis=1.3.7=h1a28f6b_0
libvpx=1.13.1=h313beb8_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_1
libxml2=2.10.4=h0b34f26_2
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_1
markupsafe=2.1.3=py310h80987f9_0
mathjax=2.7.7=hce30654_3
matplotlib=3.9.2=pypi_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.3.0=py310hca03da5_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=py310hca03da5_0
numpy=2.0.1=py310h3b2db8e_1
numpy-base=2.0.1=py310ha9811e2_1
omegaconf=2.3.0=pypi_0
open3d=0.18.0=pypi_0
openh264=2.1.1=h98b2900_0
openjpeg=2.5.2=h54b8e55_0
openssl=3.4.0=h39f12f2_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pango=1.50.7=h7271ec9_0
pcre2=10.42=hb066dcc_1
pillow=11.0.0=py310hfaf4e14_0
pip=24.2=py310hca03da5_0
pixman=0.40.0=h1a28f6b_0
platformdirs=4.3.6=pypi_0
plotly=5.24.1=pypi_0
portalocker=2.3.0=py310hca03da5_1
protobuf=5.28.3=pypi_0
psutil=6.1.0=pypi_0
pyparsing=3.2.0=pypi_0
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.15=hb885b13_1
python-dateutil=2.9.0.post0=pypi_0
python-kaleido=0.2.1=pyhd8ed1ab_0
pytorch=2.1.2=py3.10_0
pytorch3d=0.7.8=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=py310h80987f9_0
readline=8.2=h1a28f6b_0
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=py310hca03da5_1
retrying=1.3.4=pypi_0
rpds-py=0.21.0=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
seaborn=0.13.2=pypi_0
sentry-sdk=2.19.0=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=75.1.0=py310hca03da5_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.45.3=h80987f9_0
sympy=1.13.2=py310hca03da5_0
tenacity=9.0.0=pypi_0
tesseract=5.2.0=hc377ac9_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h6ba3021_0
torchaudio=2.1.2=py310_cpu
torchvision=0.16.2=py310_cpu
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
trimesh=4.5.2=pypi_0
typing-extensions=4.11.0=py310hca03da5_0
typing_extensions=4.11.0=py310hca03da5_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=py310hca03da5_0
wandb=0.18.7=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.6=pypi_0
wheel=0.44.0=py310hca03da5_0
xz=5.4.6=h80987f9_1
yaml=0.2.5=h1a28f6b_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h18a0788_1
zstandard=0.23.0=pypi_0
zstd=1.5.6=hfb09047_0
